+++ lib/src/core.dart
@@ -13,6 +13,13 @@ abstract class BaseResultOr<T, T2> {}
 /// Predefined error types are [NonFatalResultError], [FatalResultError] types.
 /// You can extend these types, or define custom base error types using [BaseResultOr]
 sealed class ResultOr<T> extends BaseResultOr<T, BaseResultError> {
+
+  /// Handy getters
+  bool get isSuccess => this is ResultData<T>;
+  bool get isError => this is ResultError<T>;
+  T? get dataOrNull => this is ResultData<T> ? (this as ResultData<T>).data : null;
+  BaseResultError? get errorOrNull => this is ResultError<T> ? (this as ResultError<T>).error : null;
+
   ResultOr();
 
   /// Function wraps any other function and return either expected value, or error class.
@@ -25,20 +32,9 @@ sealed class ResultOr<T> extends BaseResultOr<T, BaseResultError> {
       var result = ResultData<T>(data: func());
       onSuccess?.call(result.data);
       return result;
-    } on NonFatalResultError catch (e) {
-      var error = ResultError<T>(error: e);
-      onError?.call(error.error);
-      if (kDebugMode) {
-        print(e);
-      }
-      return error;
     } catch (e, s) {
-      var error = ResultError<T>(error: FatalResultError(e.toString(), s, e));
+      var error = _mapThrown<T>(e, s);
       onError?.call(error.error);
-      if (kDebugMode) {
-        print(e);
-        print(s);
-      }
       return error;
     }
   }
@@ -53,20 +49,9 @@ sealed class ResultOr<T> extends BaseResultOr<T, BaseResultError> {
       var result = ResultData<T>(data: await func());
       onSuccess?.call(result.data);
       return result;
-    } on NonFatalResultError catch (e) {
-      var error = ResultError<T>(error: e);
-      onError?.call(error.error);
-      if (kDebugMode) {
-        print(e);
-      }
-      return error;
     } catch (e, s) {
-      var error = ResultError<T>(error: FatalResultError(e.toString(), s, e));
+      var error = _mapThrown<T>(e, s);
       onError?.call(error.error);
-      if (kDebugMode) {
-        print(e);
-        print(s);
-      }
       return error;
     }
   }
@@ -76,6 +61,38 @@ sealed class ResultOr<T> extends BaseResultOr<T, BaseResultError> {
     return Stream.eventTransformed(stream.map((data) {
       return ResultData(data: data);
     }), (sink) => _ResultOrDuplicateSink(sink));
+  }
+
+  static ResultError<T> _mapThrown<T>(e, s) {
+    if (e case BaseResultError()) {
+      var error = ResultError<T>(error: e);
+      if (kDebugMode) {
+        print(e);
+        print(s);
+      }
+      return error;
+    } else if (e case Exception()) {
+      var error = ResultError<T>(error: NonFatalResultError(e.toString(), s, e));
+      if (kDebugMode) {
+        print(e);
+        print(s);
+      }
+      return error;
+    } else if (e case Error()) {
+      var error = ResultError<T>(error: FatalResultError(e.toString(), s, e));
+      if (kDebugMode) {
+        print(e);
+        print(s);
+      }
+      return error;
+    } else {
+      var error = ResultError<T>(error: UnexpectedResultError(e.toString(), s, e));
+      if (kDebugMode) {
+        print(e);
+        print(s);
+      }
+      return error;
+    }
   }
 }
 
@@ -91,17 +108,8 @@ class _ResultOrDuplicateSink<T> implements EventSink<ResultOr<T>> {
 
   @override
   void addError(Object e, [StackTrace? s]) {
-    if (e is NonFatalResultError) {
-      var error = ResultError<T>(error: e);
-      _outputSink.add(error);
-    } else {
-      var error = ResultError<T>(error: FatalResultError(e.toString(), s, e));
-      _outputSink.add(error);
-    }
-    if (kDebugMode) {
-      print(e);
-      print(s);
-    }
+    var error = ResultOr._mapThrown<T>(e, s);
+    _outputSink.add(error);
   }
 
   @override