+++ CHANGELOG.md
@@ -9,3 +9,11 @@
 ## 0.8.6
 
 * Types fixed
+
+## 0.8.7
+
+* Error types refactored
+* New getters and extensions
+* Minor refactoring and bug fixes
+* Tests commited
+* Readme updated